Extract generic web preview state dispatch
