+++ control_plane/drivers/generic_web_preview_dispatch.py
@@ -6,9 +6,14 @@ from typing import Protocol, cast
 import click
 from pydantic import BaseModel, ConfigDict, Field, field_validator, model_validator
 
+from control_plane.contracts.preview_desired_state_record import PreviewDesiredStateRecord
 from control_plane.contracts.preview_generation_record import (
     PreviewGenerationRecord,
     PreviewGenerationState,
+)
+from control_plane.contracts.preview_inventory_scan_record import (
+    PreviewInventoryScanRecord,
+    build_preview_inventory_scan_id,
 )
 from control_plane.contracts.preview_mutation_request import (
     PreviewGenerationMutationRequest,
@@ -21,6 +26,7 @@ from control_plane.drivers.dispatch import (
     _DriverRouteExecutionMetadata,
     _ProductRouteEnvelope,
     _ResolvedProductDriverContext,
+    ProductDriverMismatchError,
     _normalize_preview_verification_checked_urls,
     _normalize_release_status,
     _validate_driver_envelope_product,
@@ -29,7 +35,131 @@ from control_plane.launchplane_mutations import (
     LaunchplaneMutationStore,
     apply_launchplane_generation_evidence,
 )
+from control_plane.workflows.generic_web_deploy import product_profile_uses_generic_web_base
+from control_plane.workflows.generic_web_preview import (
+    GenericWebPreviewDesiredStateRequest,
+    GenericWebPreviewDestroyRequest,
+    GenericWebPreviewInventoryRequest,
+    GenericWebPreviewProfileStore,
+    GenericWebPreviewReadinessRequest,
+    GenericWebPreviewRefreshRequest,
+    discover_generic_web_preview_desired_state,
+    evaluate_generic_web_preview_readiness,
+    execute_generic_web_preview_inventory,
+)
 from control_plane.workflows.launchplane import find_preview_record
+from control_plane.workflows.ship import utc_now_timestamp
+
+
+class GenericWebPreviewDesiredStateEnvelope(_ProductRouteEnvelope):
+    schema_version: int = Field(default=1, ge=1)
+    desired_state: GenericWebPreviewDesiredStateRequest
+
+    @model_validator(mode="after")
+    def _validate_alignment(self) -> "GenericWebPreviewDesiredStateEnvelope":
+        if not self.product.strip():
+            raise ValueError("generic web preview desired state requires product")
+        if self.product.strip() != self.desired_state.product.strip():
+            raise ValueError("generic web preview desired state requires matching product values")
+        return self
+
+
+_GENERIC_WEB_PREVIEW_DESIRED_STATE_ROUTE = _DriverRouteExecutionMetadata(
+    route_path="/v1/drivers/generic-web/preview-desired-state",
+    envelope_model=GenericWebPreviewDesiredStateEnvelope,
+    denial_message=(
+        "Workflow cannot discover generic web preview desired state"
+        " for the requested product/context."
+    ),
+)
+
+
+class GenericWebPreviewRefreshEnvelope(_ProductRouteEnvelope):
+    schema_version: int = Field(default=1, ge=1)
+    refresh: GenericWebPreviewRefreshRequest
+
+    @model_validator(mode="after")
+    def _validate_alignment(self) -> "GenericWebPreviewRefreshEnvelope":
+        if not self.product.strip():
+            raise ValueError("generic web preview refresh requires product")
+        if self.product.strip() != self.refresh.product.strip():
+            raise ValueError("generic web preview refresh requires matching product values")
+        return self
+
+
+class GenericWebPreviewInventoryEnvelope(_ProductRouteEnvelope):
+    schema_version: int = Field(default=1, ge=1)
+    inventory: GenericWebPreviewInventoryRequest
+
+    @model_validator(mode="after")
+    def _validate_alignment(self) -> "GenericWebPreviewInventoryEnvelope":
+        if not self.product.strip():
+            raise ValueError("generic web preview inventory requires product")
+        if self.product.strip() != self.inventory.product.strip():
+            raise ValueError("generic web preview inventory requires matching product values")
+        return self
+
+
+_GENERIC_WEB_PREVIEW_REFRESH_ROUTE = _DriverRouteExecutionMetadata(
+    route_path="/v1/drivers/generic-web/preview-refresh",
+    envelope_model=GenericWebPreviewRefreshEnvelope,
+    denial_message=(
+        "Workflow cannot refresh generic web preview state for the requested product/context."
+    ),
+)
+
+
+_GENERIC_WEB_PREVIEW_INVENTORY_ROUTE = _DriverRouteExecutionMetadata(
+    route_path="/v1/drivers/generic-web/preview-inventory",
+    envelope_model=GenericWebPreviewInventoryEnvelope,
+    denial_message=(
+        "Workflow cannot read generic web preview inventory for the requested product/context."
+    ),
+)
+
+
+class GenericWebPreviewReadinessEnvelope(_ProductRouteEnvelope):
+    schema_version: int = Field(default=1, ge=1)
+    readiness: GenericWebPreviewReadinessRequest
+
+    @model_validator(mode="after")
+    def _validate_alignment(self) -> "GenericWebPreviewReadinessEnvelope":
+        if not self.product.strip():
+            raise ValueError("generic web preview readiness requires product")
+        if self.product.strip() != self.readiness.product.strip():
+            raise ValueError("generic web preview readiness requires matching product values")
+        return self
+
+
+_GENERIC_WEB_PREVIEW_READINESS_ROUTE = _DriverRouteExecutionMetadata(
+    route_path="/v1/drivers/generic-web/preview-readiness",
+    envelope_model=GenericWebPreviewReadinessEnvelope,
+    denial_message=(
+        "Workflow cannot evaluate generic web preview readiness for the requested product/context."
+    ),
+)
+
+
+class GenericWebPreviewDestroyEnvelope(_ProductRouteEnvelope):
+    schema_version: int = Field(default=1, ge=1)
+    destroy: GenericWebPreviewDestroyRequest
+
+    @model_validator(mode="after")
+    def _validate_alignment(self) -> "GenericWebPreviewDestroyEnvelope":
+        if not self.product.strip():
+            raise ValueError("generic web preview destroy requires product")
+        if self.product.strip() != self.destroy.product.strip():
+            raise ValueError("generic web preview destroy requires matching product values")
+        return self
+
+
+_GENERIC_WEB_PREVIEW_DESTROY_ROUTE = _DriverRouteExecutionMetadata(
+    route_path="/v1/drivers/generic-web/preview-destroy",
+    envelope_model=GenericWebPreviewDestroyEnvelope,
+    denial_message=(
+        "Workflow cannot destroy generic web preview state for the requested product/context."
+    ),
+)
 
 
 class GenericWebPreviewVerificationRequest(BaseModel):
@@ -140,6 +270,134 @@ def _handle_generic_web_preview_verification(
             request=request.verification,
         )
     )
+
+
+def _validate_generic_web_preview_profile(
+    request: _ProductRouteEnvelope,
+    resolved_context: _ResolvedProductDriverContext,
+    record_store: object,
+    control_plane_root_path: Path,
+) -> None:
+    del request, record_store, control_plane_root_path
+    profile = resolved_context.profile
+    if profile is None:
+        raise ProductDriverMismatchError("Generic web preview routes require a product profile.")
+    if not product_profile_uses_generic_web_base(profile):
+        raise click.ClickException(
+            f"Product {profile.product!r} is configured for driver {profile.driver_id!r}, "
+            "not generic-web or a generic-web based driver."
+        )
+    if not profile.preview.enabled:
+        raise click.ClickException(
+            f"Product {profile.product!r} does not have generic-web previews enabled."
+        )
+    if not profile.preview.context.strip():
+        raise click.ClickException(
+            f"Product {profile.product!r} does not define a preview context."
+        )
+
+
+def _write_preview_inventory_scan_if_supported(
+    *,
+    record_store: object,
+    context: str,
+    source: str,
+    preview_slugs: tuple[str, ...],
+) -> str:
+    if not hasattr(record_store, "write_preview_inventory_scan_record"):
+        return ""
+    scanned_at = utc_now_timestamp()
+    scan_id = build_preview_inventory_scan_id(
+        context_name=context,
+        scanned_at=scanned_at,
+    )
+    getattr(record_store, "write_preview_inventory_scan_record")(
+        PreviewInventoryScanRecord(
+            scan_id=scan_id,
+            context=context,
+            scanned_at=scanned_at,
+            source=source,
+            status="pass",
+            preview_count=len(preview_slugs),
+            preview_slugs=preview_slugs,
+        )
+    )
+    return scan_id
+
+
+def _write_preview_desired_state_if_supported(
+    *, record_store: object, record: PreviewDesiredStateRecord
+) -> str:
+    if not hasattr(record_store, "write_preview_desired_state_record"):
+        return ""
+    getattr(record_store, "write_preview_desired_state_record")(record)
+    return record.desired_state_id
+
+
+def _handle_generic_web_preview_desired_state(
+    request: GenericWebPreviewDesiredStateEnvelope,
+    resolved_context: _ResolvedProductDriverContext,
+    record_store: object,
+    control_plane_root_path: Path,
+) -> _DescriptorDriverDispatchResult:
+    assert resolved_context.profile is not None
+    driver_result = discover_generic_web_preview_desired_state(
+        control_plane_root=control_plane_root_path,
+        record_store=cast(GenericWebPreviewProfileStore, record_store),
+        request=request.desired_state,
+        discovered_at=utc_now_timestamp(),
+        profile=resolved_context.profile,
+    )
+    preview_desired_state_id = _write_preview_desired_state_if_supported(
+        record_store=record_store,
+        record=driver_result,
+    )
+    return _DescriptorDriverDispatchResult(
+        result={"preview_desired_state_id": preview_desired_state_id},
+        driver_result=driver_result,
+    )
+
+
+def _handle_generic_web_preview_inventory(
+    request: GenericWebPreviewInventoryEnvelope,
+    resolved_context: _ResolvedProductDriverContext,
+    record_store: object,
+    control_plane_root_path: Path,
+) -> _DescriptorDriverDispatchResult:
+    assert resolved_context.profile is not None
+    driver_result = execute_generic_web_preview_inventory(
+        control_plane_root=control_plane_root_path,
+        record_store=cast(GenericWebPreviewProfileStore, record_store),
+        request=request.inventory,
+        profile=resolved_context.profile,
+    )
+    preview_inventory_scan_id = _write_preview_inventory_scan_if_supported(
+        record_store=record_store,
+        context=driver_result.context,
+        source=driver_result.source,
+        preview_slugs=tuple(item.previewSlug for item in driver_result.previews),
+    )
+    return _DescriptorDriverDispatchResult(
+        result={"preview_inventory_scan_id": preview_inventory_scan_id},
+        driver_result=driver_result,
+    )
+
+
+def _handle_generic_web_preview_readiness(
+    request: GenericWebPreviewReadinessEnvelope,
+    resolved_context: _ResolvedProductDriverContext,
+    record_store: object,
+    control_plane_root_path: Path,
+) -> _DescriptorDriverDispatchResult:
+    assert resolved_context.profile is not None
+    driver_result = evaluate_generic_web_preview_readiness(
+        control_plane_root=control_plane_root_path,
+        record_store=cast(GenericWebPreviewProfileStore, record_store),
+        request=request.readiness,
+        checked_at=utc_now_timestamp(),
+        profile=resolved_context.profile,
+    )
+    return _DescriptorDriverDispatchResult(result={}, driver_result=driver_result)
 
 
 def _apply_generic_web_preview_verification_records(
